# docs/openapi/audit-service.json
{"openapi":"3.1.0","info":{"title":"Audit Service","description":"Activity logs, compliance events, and data access tracking","version":"1.0.0"},"paths":{"/health":{"get":{"summary":"Health Check","description":"Health check endpoint.","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/":{"get":{"summary":"Root","description":"Root endpoint.","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/audit/logs":{"post":{"tags":["audit-logs"],"summary":"Create Audit Log","description":"Create a new audit log entry.","operationId":"create_audit_log_api_v1_audit_logs_post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuditLogCreate"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuditLogResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["audit-logs"],"summary":"List Audit Logs","description":"List audit logs with optional filters.","operationId":"list_audit_logs_api_v1_audit_logs_get","parameters":[{"name":"actor_id","in":"query","required":false,"schema":{"anyOf":[{"type":"string","format":"uuid"},{"type":"null"}],"title":"Actor Id"}},{"name":"action","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Action"}},{"name":"resource_type","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Resource Type"}},{"name":"resource_id","in":"query","required":false,"schema":{"anyOf":[{"type":"string","format":"uuid"},{"type":"null"}],"title":"Resource Id"}},{"name":"service","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Service"}},{"name":"status","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Status"}},{"name":"skip","in":"query","required":false,"schema":{"type":"integer","minimum":0,"default":0,"title":"Skip"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"default":20,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response List Audit Logs Api V1 Audit Logs Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/audit/logs/{log_id}":{"get":{"tags":["audit-logs"],"summary":"Get Audit Log","description":"Get audit log by ID.","operationId":"get_audit_log_api_v1_audit_logs__log_id__get","parameters":[{"name":"log_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Log Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuditLogResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/audit/resource/{resource_id}":{"get":{"tags":["audit-logs"],"summary":"Get Resource Audit History","description":"Get complete audit history for a resource.","operationId":"get_resource_audit_history_api_v1_resource__resource_id__get","parameters":[{"name":"resource_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Resource Id"}},{"name":"resource_type","in":"query","required":true,"schema":{"type":"string","title":"Resource Type"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/audit/user/{actor_id}/activity":{"get":{"tags":["audit-logs"],"summary":"Get User Activity","description":"Get recent activity for a user.","operationId":"get_user_activity_api_v1_user__actor_id__activity_get","parameters":[{"name":"actor_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Actor Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/audit/stats":{"get":{"tags":["audit-logs"],"summary":"Get Audit Stats","description":"Get audit statistics.","operationId":"get_audit_stats_api_v1audit_stats_get","parameters":[{"name":"days","in":"query","required":false,"schema":{"type":"integer","maximum":365,"minimum":1,"default":30,"title":"Days"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/audit/compliance":{"post":{"tags":["compliance"],"summary":"Create Compliance Event","description":"Create a new compliance event.","operationId":"create_compliance_event_api_v1_compliance_post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ComplianceEventCreate"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ComplianceEventResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["compliance"],"summary":"List Compliance Events","description":"List compliance events.","operationId":"list_compliance_events_api_v1_compliance_get","parameters":[{"name":"severity","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Severity"}},{"name":"event_type","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Event Type"}},{"name":"skip","in":"query","required":false,"schema":{"type":"integer","minimum":0,"default":0,"title":"Skip"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"default":50,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response List Compliance Events Api V1 Compliance Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/audit/compliance/{event_id}":{"get":{"tags":["compliance"],"summary":"Get Compliance Event","description":"Get compliance event by ID.","operationId":"get_compliance_event_api_v1_compliance__event_id__get","parameters":[{"name":"event_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Event Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ComplianceEventResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/audit/data-access":{"post":{"tags":["data-access"],"summary":"Log Data Access","description":"Log a data access event.","operationId":"log_data_access_api_v1_data_access_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DataAccessLogCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DataAccessLogResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/audit/data-access/unauthorized":{"get":{"tags":["data-access"],"summary":"Get Unauthorized Accesses","description":"Get unauthorized data access attempts.","operationId":"get_unauthorized_accesses_api_v1_data_access_unauthorized_get","parameters":[{"name":"days","in":"query","required":false,"schema":{"type":"integer","maximum":365,"minimum":1,"default":30,"title":"Days"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":500,"minimum":1,"default":100,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/audit/data-access/summary":{"get":{"tags":["data-access"],"summary":"Get Data Access Summary","description":"Get data access summary.","operationId":"get_data_access_summary_api_v1_data_access_summary_get","parameters":[{"name":"days","in":"query","required":false,"schema":{"type":"integer","maximum":365,"minimum":1,"default":30,"title":"Days"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DataAccessSummary"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/audit/data-access/{log_id}":{"get":{"tags":["data-access"],"summary":"Get Data Access Log","description":"Get data access log by ID.","operationId":"get_data_access_log_api_v1_data_access__log_id__get","parameters":[{"name":"log_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Log Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DataAccessLogResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"AccessMethodEnum":{"type":"string","enum":["api","batch","internal","export"],"title":"AccessMethodEnum","description":"Data access methods."},"AuditLogCreate":{"properties":{"actor_id":{"type":"string","format":"uuid","title":"Actor Id"},"action":{"type":"string","title":"Action"},"resource_id":{"type":"string","format":"uuid","title":"Resource Id"},"resource_type":{"type":"string","title":"Resource Type"},"old_values":{"anyOf":[{"type":"object"},{"type":"null"}],"title":"Old Values"},"new_values":{"anyOf":[{"type":"object"},{"type":"null"}],"title":"New Values"},"changes":{"anyOf":[{"type":"object"},{"type":"null"}],"title":"Changes"},"status":{"allOf":[{"$ref":"#/components/schemas/AuditStatusEnum"}],"default":"success"},"error_message":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Error Message"},"ip_address":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Ip Address"},"user_agent":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"User Agent"},"service":{"type":"string","title":"Service"},"project_id":{"anyOf":[{"type":"string","format":"uuid"},{"type":"null"}],"title":"Project Id"},"issue_id":{"anyOf":[{"type":"string","format":"uuid"},{"type":"null"}],"title":"Issue Id"}},"type":"object","required":["actor_id","action","resource_id","resource_type","service"],"title":"AuditLogCreate","description":"Create audit log."},"AuditLogResponse":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"actor_id":{"type":"string","format":"uuid","title":"Actor Id"},"action":{"type":"string","title":"Action"},"resource_id":{"type":"string","format":"uuid","title":"Resource Id"},"resource_type":{"type":"string","title":"Resource Type"},"old_values":{"anyOf":[{"type":"object"},{"type":"null"}],"title":"Old Values"},"new_values":{"anyOf":[{"type":"object"},{"type":"null"}],"title":"New Values"},"changes":{"anyOf":[{"type":"object"},{"type":"null"}],"title":"Changes"},"status":{"type":"string","title":"Status"},"error_message":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Error Message"},"ip_address":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Ip Address"},"user_agent":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"User Agent"},"service":{"type":"string","title":"Service"},"project_id":{"anyOf":[{"type":"string","format":"uuid"},{"type":"null"}],"title":"Project Id"},"issue_id":{"anyOf":[{"type":"string","format":"uuid"},{"type":"null"}],"title":"Issue Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["id","actor_id","action","resource_id","resource_type","status","service","created_at"],"title":"AuditLogResponse","description":"Audit log response."},"AuditStatusEnum":{"type":"string","enum":["success","failure","warning"],"title":"AuditStatusEnum","description":"Audit status types."},"ComplianceEventCreate":{"properties":{"event_type":{"type":"string","title":"Event Type"},"severity":{"$ref":"#/components/schemas/SeverityEnum"},"actor_id":{"type":"string","format":"uuid","title":"Actor Id"},"action":{"type":"string","title":"Action"},"resource_id":{"anyOf":[{"type":"string","format":"uuid"},{"type":"null"}],"title":"Resource Id"},"resource_type":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Resource Type"},"description":{"type":"string","title":"Description"},"details":{"anyOf":[{"type":"object"},{"type":"null"}],"title":"Details"},"retention_days":{"type":"string","title":"Retention Days","default":"permanent"}},"type":"object","required":["event_type","severity","actor_id","action","description"],"title":"ComplianceEventCreate","description":"Create compliance event."},"ComplianceEventResponse":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"event_type":{"type":"string","title":"Event Type"},"severity":{"type":"string","title":"Severity"},"actor_id":{"type":"string","format":"uuid","title":"Actor Id"},"action":{"type":"string","title":"Action"},"resource_id":{"anyOf":[{"type":"string","format":"uuid"},{"type":"null"}],"title":"Resource Id"},"resource_type":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Resource Type"},"description":{"type":"string","title":"Description"},"details":{"anyOf":[{"type":"object"},{"type":"null"}],"title":"Details"},"retention_days":{"type":"string","title":"Retention Days"},"is_deleted":{"type":"boolean","title":"Is Deleted"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"expires_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Expires At"}},"type":"object","required":["id","event_type","severity","actor_id","action","description","retention_days","is_deleted","created_at"],"title":"ComplianceEventResponse","description":"Compliance event response."},"DataAccessLogCreate":{"properties":{"actor_id":{"type":"string","format":"uuid","title":"Actor Id"},"resource_id":{"type":"string","format":"uuid","title":"Resource Id"},"resource_type":{"type":"string","title":"Resource Type"},"fields_accessed":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Fields Accessed"},"access_method":{"$ref":"#/components/schemas/AccessMethodEnum"},"operation":{"$ref":"#/components/schemas/OperationEnum"},"ip_address":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Ip Address"},"purpose":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Purpose"},"was_authorized":{"type":"boolean","title":"Was Authorized","default":true}},"type":"object","required":["actor_id","resource_id","resource_type","access_method","operation"],"title":"DataAccessLogCreate","description":"Create data access log."},"DataAccessLogResponse":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"actor_id":{"type":"string","format":"uuid","title":"Actor Id"},"resource_id":{"type":"string","format":"uuid","title":"Resource Id"},"resource_type":{"type":"string","title":"Resource Type"},"fields_accessed":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Fields Accessed"},"access_method":{"type":"string","title":"Access Method"},"operation":{"type":"string","title":"Operation"},"ip_address":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Ip Address"},"purpose":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Purpose"},"was_authorized":{"type":"boolean","title":"Was Authorized"},"accessed_at":{"type":"string","format":"date-time","title":"Accessed At"}},"type":"object","required":["id","actor_id","resource_id","resource_type","access_method","operation","was_authorized","accessed_at"],"title":"DataAccessLogResponse","description":"Data access log response."},"DataAccessSummary":{"properties":{"total_accesses":{"type":"integer","title":"Total Accesses"},"unauthorized_accesses":{"type":"integer","title":"Unauthorized Accesses"},"accesses_by_method":{"additionalProperties":{"type":"integer"},"type":"object","title":"Accesses By Method"},"accesses_by_operation":{"additionalProperties":{"type":"integer"},"type":"object","title":"Accesses By Operation"},"accesses_by_actor":{"additionalProperties":{"type":"integer"},"type":"object","title":"Accesses By Actor"}},"type":"object","required":["total_accesses","unauthorized_accesses","accesses_by_method","accesses_by_operation","accesses_by_actor"],"title":"DataAccessSummary","description":"Summary of data access for compliance."},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"OperationEnum":{"type":"string","enum":["read","update","delete","export"],"title":"OperationEnum","description":"Operations for data access."},"SeverityEnum":{"type":"string","enum":["critical","high","medium","low","info"],"title":"SeverityEnum","description":"Severity levels for compliance events."},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}